backend/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
cssselect2==0.7.0
Django==4.1.7
django-cors-headers==3.14.0
django-rest-framework==0.1.0
django-stubs==1.15.0
django-stubs-ext==0.7.0
djangorestframework==3.14.0
djangorestframework-stubs==1.9.1
drf-yasg==1.21.5
factory-boy==3.2.1
Faker==17.6.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.2
mypy==1.0.1
mypy-extensions==1.0.0
oscrypto==1.3.0
packaging==23.0
Pillow==9.4.0
pycparser==2.21
pyHanko==0.17.0
pyhanko-certvalidator==0.20.1
pypdf==3.5.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.12
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
tomli==2.0.1
types-pytz==2022.7.1.2
types-PyYAML==6.0.12.8
types-requests==2.28.11.15
types-urllib3==1.26.25.8
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.14
webencodings==0.5.1
whitenoise==6.4.0
xhtml2pdf==0.2.9
